requirements.txt
sklearn
numpy
pandas
plotly.express
dash_core_components
dash_html_components
dash
lightgbm
gunicorn
requests
unidecode
pickle-mixin
pyarrow